Conflicts:
	CHANGELOG
	RELEASE_NOTES
	cocos2d/CCRibbon.m

Former-commit-id: ee88b35214031878148332798cf6c223f5dd57c6

+++ cocos2d/CCRibbon.h
@@ -115,8 +115,8 @@
 	GLubyte	colors[kCCRibbon_Max*8];
 	float		creationTime[kCCRibbon_Max];
 	BOOL		finished;
-	uint		end;
-	uint		begin;
+	NSUInteger		end;
+	NSUInteger		begin;
 }
 -(id)init;
 -(void)reset;
